Update vasp_driver to allow for persistent connections.  Needs two patches to vasp, currently tested on vasp 5.2.12 only
git-svn-id: svn+ssh://cvs/home/jrk33/repo/trunk/QUIP@10026 9aeadc09-fa51-0410-b5d8-e5c65b1e449c

+++ QUIP_FilePot_Drivers/vasp_driver.f95
@@ -65,9 +65,18 @@ subroutine do_vasp_calc(at, args_str, error)
 
    character(len=STRING_LENGTH) :: incar_template_file, kpoints_file, potcar_files, vasp_path, run_suffix, verbosity_str
    character(len=STRING_LENGTH) :: calc_energy, calc_force, calc_virial, calc_local_energy
-   logical :: do_calc_energy, do_calc_force, do_calc_virial, do_calc_local_energy
+   logical :: do_calc_energy, do_calc_force, do_calc_virial, do_calc_local_energy, persistent
    logical :: clean_up_files, ignore_convergence, no_use_WAVECAR, force_constant_basis
+   integer :: persistent_restart_interval
+   real(dp) :: persistent_max_wait_time
    type(Dictionary) :: incar_dict, cli
+
+   integer :: persistent_run_i
+   logical :: persistent_already_started = .false., persistent_calc_done
+   type(inoutput) :: persistent_io
+   real(dp) :: wait_time
+
+   integer :: run_dir_i, force_run_dir_i
 
    character(len=STRING_LENGTH) :: run_dir
    type(Inoutput) :: io
@@ -79,20 +88,23 @@ subroutine do_vasp_calc(at, args_str, error)
    INIT_ERROR(error)
 
    call initialise(cli)
-   call param_register(cli, "verbosity", "NORMAL", verbosity_str, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "INCAR_template", "INCAR.template", incar_template_file, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "kpoints_file", "KPOINTS", kpoints_file, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "potcar_files", PARAM_MANDATORY, potcar_files, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "vasp", PARAM_MANDATORY, vasp_path, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "energy", "", calc_energy, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "force", "", calc_force, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "virial", "", calc_virial, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "local_energy", "", calc_local_energy, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "clean_up_files", "T", clean_up_files, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "ignore_convergence", "F", ignore_convergence, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "no_use_WAVECAR", "F", no_use_WAVECAR, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "force_constant_basis", "F", force_constant_basis, help_string="No help yet.  This source file was $LastChangedBy$")
-   call param_register(cli, "run_suffix", "run", run_suffix, help_string="No help yet.  This source file was $LastChangedBy$")
+   call param_register(cli, "verbosity", "NORMAL", verbosity_str, help_string="verbosity level")
+   call param_register(cli, "INCAR_template", "INCAR.template", incar_template_file, help_string="name of file with template for main input (INCAR) file")
+   call param_register(cli, "kpoints_file", "KPOINTS", kpoints_file, help_string="name of k-points file (KPOINTS)")
+   call param_register(cli, "potcar_files", PARAM_MANDATORY, potcar_files, help_string="mapping from atomic numbers to POTCAR files, Z_1 POTCAR_FILE_1 [ Z_2 POTCAR_FILE_2 ... ]")
+   call param_register(cli, "vasp", PARAM_MANDATORY, vasp_path, help_string="vasp executable")
+   call param_register(cli, "energy", "", calc_energy, help_string="if set, calculate energy and put it in parameter named this")
+   call param_register(cli, "force", "", calc_force, help_string="if set, calculate forces and put them in property named this")
+   call param_register(cli, "virial", "", calc_virial, help_string="if set, calculate virial and put it in parameter named this")
+   call param_register(cli, "local_energy", "", calc_local_energy, help_string="not supported")
+   call param_register(cli, "clean_up_files", "T", clean_up_files, help_string="if true, delete run directory when done")
+   call param_register(cli, "ignore_convergence", "F", ignore_convergence, help_string="If true, accept result even if failed to converge SCF")
+   call param_register(cli, "no_use_WAVECAR", "F", no_use_WAVECAR, help_string="If true, don't reuse previous wavefunction file WAVECAR")
+   call param_register(cli, "force_constant_basis", "F", force_constant_basis, help_string="If true, do a restart with the same basis")
+   call param_register(cli, "run_suffix", "run", run_suffix, help_string="suffix to label persistent state from run to run")
+   call param_register(cli, "persistent", "F", persistent, help_string="If true, use persistent connection to a long-running vasp process")
+   call param_register(cli, "persistent_restart_interval", "100", persistent_restart_interval, help_string="If persistent, restart vasp this often")
+   call param_register(cli, "persistent_max_wait_time", "600", persistent_max_wait_time, help_string="If persistent, never wait more than this long (s) for calculation to finish")
    if (.not. param_read_line(cli, args_str, ignore_unknown=.true.,task='do_vasp_calc args_str')) then
       call print("Args:  verbosity INCAR_template kpoints_file potcar_files vasp energy force", PRINT_ALWAYS)
       call print("       virial local_energy clean_up_files ignore_convergence no_use_WAVECAR", PRINT_ALWAYS)
@@ -122,6 +134,11 @@ subroutine do_vasp_calc(at, args_str, error)
    call print("  no_use_WAVECAR " // no_use_WAVECAR, PRINT_VERBOSE)
    call print("  force_constant_basis " // force_constant_basis, PRINT_VERBOSE)
    call print("  run_suffix " // run_suffix, PRINT_VERBOSE)
+   call print("  persistent " // persistent, PRINT_VERBOSE)
+   if (persistent) then
+      call print("  persistent_restart_interval " // persistent_restart_interval, PRINT_VERBOSE)
+      call print("  persistent_max_wait_time " // persistent_max_wait_time, PRINT_VERBOSE)
+   endif
 
    ! check for local energy
    if (do_calc_local_energy) then
@@ -137,17 +154,35 @@ subroutine do_vasp_calc(at, args_str, error)
    ! check for stress, set ISIF
    if (.not. do_calc_energy .and. .not. do_calc_force .and. .not. do_calc_virial) then
       return
-   else if (.not. do_calc_virial) then
+   else if (.not. persistent .and. .not. do_calc_virial) then
       call set_value(incar_dict, "ISIF", 0)
    else
       call set_value(incar_dict, "ISIF", 2)
    endif
 
-   run_dir = make_run_directory("vasp_run")
+   force_run_dir_i = -1
+   if (persistent) then ! make run_dir manually
+      inquire(file="persistent_run_i", exist=persistent_already_started)
+      run_dir = "vasp_run_0"
+      force_run_dir_i = 0
+      if (.not. persistent_already_started) then
+	 call system_command("mkdir "//trim(run_dir))
+	 call initialise(persistent_io, "persistent_run_i", action=OUTPUT)
+	 call print("-1", file=persistent_io)
+	 call finalise(persistent_io)
+      endif
+   endif
+   run_dir = make_run_directory("vasp_run", force_run_dir_i, run_dir_i)
 
+   if (persistent) then
+   ! set dynamics to REFTRAJ, number of ionic steps to 1000000
+      call set_value(incar_dict, "NSW", 1000000)
+      call set_value(incar_dict, "IBRION", 12)
+   else
    ! set dynamics to MD, number of ionic steps to 0
-   call set_value(incar_dict, "NSW", 0)
-   call set_value(incar_dict, "IBRION", 0)
+      call set_value(incar_dict, "NSW", 0)
+      call set_value(incar_dict, "IBRION", 0)
+   endif
    if (no_use_WAVECAR) then
       call set_value(incar_dict, "ISTART", 0)
    else
@@ -166,19 +201,21 @@ subroutine do_vasp_calc(at, args_str, error)
       endif
    endif
 
-   ! write INCAR
-   call print("writing INCAR", PRINT_VERBOSE)
-   call initialise(io, trim(run_dir)//"/INCAR", action=OUTPUT, error=error)
-   PASS_ERROR(error)
-   call print(write_string(incar_dict, entry_sep=quip_new_line), file=io)
-   call finalise(io)
+   if (.not. persistent_already_started) then
+      ! write INCAR
+      call print("writing INCAR", PRINT_VERBOSE)
+      call initialise(io, trim(run_dir)//"/INCAR", action=OUTPUT, error=error)
+      PASS_ERROR(error)
+      call print(write_string(incar_dict, entry_sep=quip_new_line), file=io)
+      call finalise(io)
 
-   ! write KPOINTS if needed
-   if (trim(kpoints_file) /= "_NONE_") then
-      call print("writing KPOINTS", PRINT_VERBOSE)
-      call system_command("cp "//trim(kpoints_file)//" "//trim(run_dir)//"/KPOINTS", status=stat)
-      if (stat /= 0) then
-	 RAISE_ERROR("do_vasp_calc copying kpoints_file "//trim(kpoints_file)//" into "//trim(run_dir), error)
+      ! write KPOINTS if needed
+      if (trim(kpoints_file) /= "_NONE_") then
+	 call print("writing KPOINTS", PRINT_VERBOSE)
+	 call system_command("cp "//trim(kpoints_file)//" "//trim(run_dir)//"/KPOINTS", status=stat)
+	 if (stat /= 0) then
+	    RAISE_ERROR("do_vasp_calc copying kpoints_file "//trim(kpoints_file)//" into "//trim(run_dir), error)
+	 endif
       endif
    endif
 
@@ -191,28 +228,116 @@ subroutine do_vasp_calc(at, args_str, error)
    call atoms_sort(at, 'Z', error=error)
    PASS_ERROR(error)
 
-   call print("writing POTCAR", PRINT_VERBOSE)
-   call write_vasp_potcar(at, trim(run_dir), trim(potcar_files), error=error)
-   PASS_ERROR(error)
-   call print("writing POSCAR", PRINT_VERBOSE)
-   call write_vasp_poscar(at, trim(run_dir), error=error)
-   PASS_ERROR(error)
-
-   call print("running vasp", PRINT_VERBOSE)
-   call print("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" >> ../vasp.stdout 2>&1")
-   call system_command("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" >> ../vasp.stdout 2>&1", status=stat)
-   if (stat /= 0) then
-      RAISE_ERROR("error running "//trim(vasp_path)//", status="//stat, error)
+   if (persistent) then
+      call print("reading persistent_run_i", PRINT_VERBOSE)
+      call initialise(persistent_io, "persistent_run_i", action=INPUT)
+      line = read_line(persistent_io)
+      read (unit=line, fmt=*, iostat=stat) persistent_run_i
+      call finalise(persistent_io)
+      call print("got persistent_run_i="//persistent_run_i, PRINT_VERBOSE)
+      persistent_run_i = persistent_run_i + 1
+      call initialise(persistent_io, "persistent_run_i", action=OUTPUT)
+      call print(persistent_run_i, file=persistent_io)
+      call finalise(persistent_io)
    endif
 
-   call system_command("fgrep -i 'total cpu time used' "//trim(run_dir)//"/OUTCAR; fgrep -c 'LOOP:' "//trim(run_dir)//"/OUTCAR")
+   if (.not. persistent_already_started) then
+      call print("writing POTCAR", PRINT_VERBOSE)
+      call write_vasp_potcar(at, trim(run_dir), trim(potcar_files), error=error)
+      PASS_ERROR(error)
+   endif
 
-   call print("reading vasp output", PRINT_VERBOSE)
-   call read_vasp_output(trim(run_dir), do_calc_energy, do_calc_force, do_calc_virial, converged, error=error)
-   PASS_ERROR(error)
+   if (.not. persistent_already_started .or. mod(persistent_run_i, persistent_restart_interval) == 0) then
+      call print("writing POSCAR", PRINT_VERBOSE)
+      call write_vasp_poscar(at, trim(run_dir), error=error)
+      PASS_ERROR(error)
+   endif
 
-   if (.not. ignore_convergence .and. .not. converged) then
-      RAISE_ERROR("failed to find sign of convergence in OUTCAR", error)
+   if (persistent) then ! restart VASP if needed, write to REFTRAJCAR if needed, initiate calc
+      if (mod(persistent_run_i, persistent_restart_interval) == 0) then ! start or restart VASP
+	 call print("starting or restarting VASP", PRINT_VERBOSE)
+	 if (persistent_already_started) then ! tell currently running vasp to quit
+	    call initialise(persistent_io, trim(run_dir)//"/REFTRAJCAR", action=OUTPUT)
+	    call print("0", file=persistent_io)
+	    call finalise(persistent_io)
+	    call initialise(persistent_io, trim(run_dir)//"/REFTRAJ_READY", action=OUTPUT)
+	    call print("-", file=persistent_io)
+	    call finalise(persistent_io)
+	    call fusleep(5000000)
+	    call system_command("cd "//trim(run_dir)//"; rm -f REFTRAJCAR REFTRAJ_READY OUTCAR CONTCAR OSZICAR")
+	    call system_command("cat vasp.stdout."//run_dir_i//" >> vasp.stdout")
+	 endif
+	 ! start vasp in background
+	 call print("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" > ../vasp.stdout."//run_dir_i//" 2>&1 &")
+	 call system_command("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" > ../vasp.stdout."//run_dir_i//" 2>&1 &", status=stat)
+	 if (stat /= 0) then
+	    RAISE_ERROR("error running "//trim(vasp_path)//", status="//stat, error)
+	 endif
+	 persistent_already_started = .true.
+      endif
+      if (mod(persistent_run_i,persistent_restart_interval) /= 0) then ! need to write current coords to REFTRAJCAR
+	 call print("writing current coords to REFTRAJCAR", PRINT_VERBOSE)
+	 call initialise(persistent_io, trim(run_dir)//"/REFTRAJCAR", action=OUTPUT)
+	 call print(at%N, file=persistent_io)
+	 call print(at%lattice(:,1), file=persistent_io)
+	 call print(at%lattice(:,2), file=persistent_io)
+	 call print(at%lattice(:,3), file=persistent_io)
+	 do i=1, at%N
+	    call print(matmul(at%g, at%pos(:,i)), file=persistent_io)
+	 end do
+	 call finalise(persistent_io)
+	 call print("touching REFTRAJ_READY", PRINT_VERBOSE)
+	 call initialise(persistent_io, trim(run_dir)//"/REFTRAJ_READY", action=OUTPUT)
+	 call print("-", file=persistent_io)
+	 call finalise(persistent_io)
+      endif
+      call print("waiting for output REFTRAJ_STEP_DONE", PRINT_VERBOSE)
+      ! wait for output to be ready
+      inquire(file=trim(run_dir)//"/REFTRAJ_STEP_DONE", exist=persistent_calc_done)
+      wait_time = 0.0_dp
+      do while (.not. persistent_calc_done)
+	 call fusleep(100000)
+	 wait_time = wait_time + 100000_dp/1.0e6_dp
+	 inquire(file=trim(run_dir)//"/REFTRAJ_STEP_DONE", exist=persistent_calc_done)
+	 if (.not. persistent_calc_done .and. wait_time > persistent_max_wait_time) then
+	    RAISE_ERROR("error waiting too long for calculation to be done", error)
+	 endif
+      end do
+   else ! not persistent, just run vasp
+      call print("running vasp", PRINT_VERBOSE)
+call system_command("cp -a "//trim(run_dir)//" "//trim(run_dir)//".pre_run")
+      call print("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" > ../vasp.stdout."//run_dir_i//" 2>&1")
+      call system_command("cd "//trim(run_dir)//"; bash -c "//'"'//trim(vasp_path)//'"'//" > ../vasp.stdout."//run_dir_i//" 2>&1", status=stat)
+      if (stat /= 0) then
+	 RAISE_ERROR("error running "//trim(vasp_path)//", status="//stat, error)
+      endif
+      call system_command("cat vasp.stdout."//run_dir_i//" >> vasp.stdout")
+   endif
+
+   ! look for errors in stdout and runtime in OUTCAR
+   call system_command("fgrep -i 'error' vasp.stdout."//run_dir_i)
+   call system_command("fgrep -i 'total cpu time used' "//trim(run_dir)//"/OUTCAR | tail -1; fgrep -c 'LOOP:' "//trim(run_dir)//"/OUTCAR | tail -1")
+
+   if (persistent) then ! read REFTRAJ output
+      call print("reading output REFTRAJ_OUTPUT", PRINT_VERBOSE)
+      ! read output
+      call read_vasp_output_persistent(at, run_dir, do_calc_energy, do_calc_force, do_calc_virial, error)
+      PASS_ERROR(error)
+      converged = .true. ! should really find a way of telling if converged for persistent connections
+      ! clean up
+      call unlink(trim(run_dir)//"/REFTRAJ_OUTPUT")
+      call unlink(trim(run_dir)//"/REFTRAJ_STEP_DONE")
+   else ! read regular VASP output (OUTCAR)
+      call print("reading vasp output", PRINT_VERBOSE)
+      call read_vasp_output(trim(run_dir), do_calc_energy, do_calc_force, do_calc_virial, converged, error=error)
+      PASS_ERROR(error)
+   endif
+
+   if (.not. converged) then
+      call print("WARNING: failed to find sign of convergence in OUTCAR", verbosity=PRINT_ALWAYS)
+      if (.not. ignore_convergence) then
+	 RAISE_ERROR("failed to find sign of convergence in OUTCAR", error)
+      endif
    endif
 
    inquire(file=trim(run_dir)//"/WAVECAR", exist=have_wavecar)
@@ -228,7 +353,7 @@ subroutine do_vasp_calc(at, args_str, error)
    call atoms_sort(at, 'sort_index', error=error)
    PASS_ERROR(error)
 
-   if (clean_up_files) then
+   if (clean_up_files .and. .not. persistent) then
       call print("cleaning up", PRINT_VERBOSE)
       call system_command("rm -rf "//trim(run_dir))
    endif
@@ -412,6 +537,54 @@ subroutine read_vasp_output(run_dir, do_calc_energy, do_calc_force, do_calc_viri
 
 end subroutine read_vasp_output
 
+subroutine read_vasp_output_persistent(at, run_dir, do_calc_energy, do_calc_force, do_calc_virial, error)
+   type(atoms), intent(inout) :: at
+   character(len=*), intent(in) :: run_dir
+   logical, intent(in) :: do_calc_energy, do_calc_force, do_calc_virial
+   integer, optional, intent(out) :: error
+
+   real(dp) :: energy, virial(3,3), force_t(3)
+   real(dp), pointer :: force_p(:,:)
+   type(inoutput) :: reftraj_output_io
+   character(len=STRING_LENGTH) :: line
+   integer :: ni
+
+   INIT_ERROR(error)
+
+   if (do_calc_force) then
+      if (.not. assign_pointer(at, "force", force_p)) then
+	 call add_property(at, "force", 0.0_dp, n_cols=3, ptr2=force_p, error=error)
+	 PASS_ERROR(error)
+      endif
+   endif
+
+   call initialise(reftraj_output_io, trim(run_dir)//"/REFTRAJ_OUTPUT", action=INPUT)
+
+   line = read_line(reftraj_output_io); read(unit=line, fmt=*) ni
+   if (ni /= at%N) then
+      RAISE_ERROR("read_vasp_output_persistent got NI == "//ni//" != at%N == "//at%N, error)
+   endif
+
+   ! read energy
+   line = read_line(reftraj_output_io); read(unit=line, fmt=*) energy
+   if (do_calc_energy) call set_param_value(at, 'energy', energy)
+   ! read force
+   do i=1, at%N
+      line = read_line(reftraj_output_io); read(unit=line, fmt=*) force_t
+      if (do_calc_force) force_p(:,i) = force_t(:)
+   end do
+   ! read virial
+   if (do_calc_virial) then
+      line = read_line(reftraj_output_io); read(unit=line, fmt=*) virial(1,1), virial(2,2), virial(3,3), virial(1,2), virial(2,3), virial(3,1)
+      virial(2,1) = virial(1,2)
+      virial(3,2) = virial(2,3)
+      virial(1,3) = virial(3,1)
+      call set_param_value(at, "virial", virial)
+   endif
+
+   call finalise(reftraj_output_io)
+end subroutine read_vasp_output_persistent
+
 subroutine read_vasp_incar_dict(incar_dict, incar_template_file, error)
    type(Dictionary), intent(inout) :: incar_dict
    character(len=*), intent(in) :: incar_template_file
@@ -421,10 +594,9 @@ subroutine read_vasp_incar_dict(incar_dict, incar_template_file, error)
    character(len=STRING_LENGTH), allocatable :: incar_a(:), incar_line_fields(:), incar_field_fields(:)
    integer :: incar_n_lines, incar_line_n_fields, incar_field_n_fields
    integer :: i, j
+   integer :: comment_pos
 
    INIT_ERROR(error)
-
-   call system_command("fgrep -i 'error' vasp.stdout")
 
    call initialise(incar_io, trim(incar_template_file), INPUT)
    call read_file(incar_io, incar_a, incar_n_lines)
@@ -435,6 +607,12 @@ subroutine read_vasp_incar_dict(incar_dict, incar_template_file, error)
    allocate(incar_field_fields(100))
    ! loop over lines
    do i=1, incar_n_lines
+      comment_pos = index(trim(incar_a(i)), '#')
+      if (comment_pos > 1) then
+	 incar_a(i) = incar_a(i)(1:comment_pos-1)
+      elseif (comment_pos == 1) then
+	 incar_a(i) = ""
+      endif
       call split_string(trim(incar_a(i)), ";", "''"//'""', incar_line_fields, incar_line_n_fields, .true.)
       ! loop over ';' separated fields
       do j=1, incar_line_n_fields
